Fix SQL integer type for crc32 field
CRC32 is an unsigned 4-byte int, and does not fit in a signed
4-byte int. There is no unsigned int type in the SQL starndard,
so use bigint instead.

+++ frontera/contrib/backends/sqlalchemy/models.py
@@ -71,7 +71,7 @@ class QueueModelMixin(object):
     score = Column(Float, index=True)
     url = Column(String(1024), nullable=False)
     fingerprint = Column(String(40), nullable=False)
-    host_crc32 = Column(Integer, nullable=False)
+    host_crc32 = Column(BigInteger, nullable=False)
     meta = Column(PickleType())
     headers = Column(PickleType())
     cookies = Column(PickleType())
